refactor: profit 패키지 vo로 수정
