+++ src/main/java/view/OutputView.java
@@ -5,7 +5,7 @@ import domain.model.Dealer;
 import domain.model.Participant;
 import domain.model.Player;
 import domain.model.Players;
-import domain.model.Profit;
+import domain.vo.Profit;
 import java.util.List;
 import java.util.StringJoiner;
 import java.util.stream.IntStream;
