+++ src/test/java/domain/model/ProfitTest.java
@@ -3,6 +3,7 @@ package domain.model;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import domain.vo.Batting;
+import domain.vo.Profit;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
