+++ src/test/java/domain/service/ProfitCalculatorTest.java
@@ -6,7 +6,7 @@ import domain.model.Card;
 import domain.model.Cards;
 import domain.model.Dealer;
 import domain.model.Player;
-import domain.model.Profit;
+import domain.vo.Profit;
 import domain.type.Letter;
 import domain.type.Suit;
 import domain.vo.Batting;
